% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{ana_build_universes}
\alias{ana_build_universes}
\title{Build universes from model scores}
\usage{
ana_build_universes(.df, .names, .definitions)
}
\arguments{
\item{.names}{vector of column names}

\item{.data}{data frame of model scores}

\item{.conditions}{vector of universe definitions}
}
\value{

}
\description{
Build universes from model scores
}
