% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{ana_calc_turnout_by_demo}
\alias{ana_calc_turnout_by_demo}
\title{Calculate expected turnout by demo}
\usage{
ana_calc_turnout_by_demo(.df, .demos, .turnout = "TurnoutGeneral")
}
\arguments{
\item{.df}{Data frame containing \code{.demos} and \code{.turnout} column}

\item{.demos}{Vector of column names containing demos}

\item{.turnout}{Name of turnout column in \code{.df}}
}
\value{
Data frame
}
\description{
Calculate expected turnout by demo
}
