% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{ana_filter_universe}
\alias{ana_filter_universe}
\title{Utility function to filter data frame for single universe}
\usage{
ana_filter_universe(.data, .universe, .target = 1)
}
\arguments{
\item{.data}{Data frame containing universe columns}

\item{.universe}{Column name of universe to use as filter}

\item{.target}{Value to filter universe for. Defaults to \code{1}}
}
\value{
Data frame
}
\description{
Utility function to filter data frame for single universe
}
\examples{
ana_filter_universe(mtcars, "vs")

}
