% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{ana_get_demo_counts}
\alias{ana_get_demo_counts}
\title{Utility function to get universe counts by demos}
\usage{
ana_get_demo_counts(.df, .demos, .universes)
}
\arguments{
\item{.df}{Data frame containing \code{.demos} and \code{.universes}}

\item{.demos}{Demo columns to group universe counts by}

\item{.universes}{Universe columns to sum}
}
\value{
Data frame
}
\description{
Utility function to get universe counts by demos
}
