% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{ana_get_demo_totals}
\alias{ana_get_demo_totals}
\title{Utility function to get count of number of voters in demo}
\usage{
ana_get_demo_totals(.df, .demos)
}
\arguments{
\item{.df}{Data frame containing \code{.demos}}

\item{.demos}{Vector of demo column names}
}
\value{
Data frame
}
\description{
Utility function to get count of number of voters in demo
}
