% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_action.R
\name{wpa_mk_12_universes}
\alias{wpa_mk_12_universes}
\title{Create 12 core action universes from a net ballot and turnout column}
\usage{
wpa_mk_12_universes(
  .data,
  .us,
  .them,
  .turnout,
  .net_hi = 0.35,
  .net_lo = 0.2,
  .to_hi = 0.9,
  .to_midhi = 0.65,
  .to_midlo = 0.5,
  .to_lo = 0.15,
  .prefixes = NULL,
  .func = "mutate"
)
}
\arguments{
\item{.data}{Data frame containing ballot and turnout columns.}

\item{.us}{Column containing ballot scores for our candidate.}

\item{.them}{Column containing ballot scores for opponent.}

\item{.turnout}{Column containing voter turnout scores.}

\item{.net_hi}{Cutoff for net ballot to assign voters in \verb{_ITB} universes.}

\item{.net_lo}{Cutoff for net ballot to assign voters in \verb{_Shore_Up} and \verb{_Marginal}
universes.}

\item{.to_hi}{Cutoff for high turnout.}

\item{.to_midhi}{Cutoff for persuasion universe.}

\item{.to_midlo}{Cutoff splitting \code{GOTV} and \code{GOTV_2}}

\item{.to_lo}{Cutoff for low turnout. Values below this are assigned to \code{Dead_Weight}.}

\item{.prefixes}{Vector of two prefixes to replace \code{R_} and \code{D_} in the core action
universes.}
}
\value{
Data frame with 12 core action universes appended.
}
\description{
Create 12 core action universes from a net ballot and turnout column
}
