% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_action.R
\name{wpa_replace_universe_prefix}
\alias{wpa_replace_universe_prefix}
\title{Replace the prefix on core action universes}
\usage{
wpa_replace_universe_prefix(.us = "R", .them = "D")
}
\arguments{
\item{.us}{Prefix to replace \code{R} in core action universes. For example, setting
\code{.us} to "Waters" replaces \code{R_ITB} and \code{R_GOTV} with \code{Waters_ITB} and \code{Waters_GOTV}.}

\item{.them}{Prefix to replace \code{D} in core action universes.}
}
\value{
Character vector
}
\description{
Replace the prefix on core action universes
}
\examples{
wpa_replace_universe_prefix(.us = "Smith", .them = "Jones")

}
